+++ digitalhouses_speedtest/rootfs/run.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/with-contenv bashio
+
 set -Eeuo pipefail
 
 bashio::log.info "Starting DigitalHouses Speedtest ${APP_VERSION:-unknown}"
@@ -9,8 +10,8 @@ export MQTT_USER="$(bashio::services mqtt username)"
 export MQTT_PASSWORD="$(bashio::services mqtt password)"
 
 if [[ -z "${MQTT_HOST}" || -z "${MQTT_PORT}" ]]; then
-  bashio::log.fatal "MQTT service information is incomplete."
-  exit 1
+    bashio::log.fatal "MQTT service information is incomplete."
+    exit 1
 fi
 
 mkdir -p /data/.config /data/runtime